Protege os endpoints PUT e DELETE com autenticação

+++ tests/test_app.py
@@ -3,6 +3,36 @@ from http import HTTPStatus
 from fast_zero.schemas import UserPublic
 
 
+def test_get_token(client, user):
+    response = client.post(
+        "/token",
+        data={"username": user.email, "password": user.clean_password},
+    )
+    token = response.json()
+
+    assert response.status_code == HTTPStatus.OK
+    assert "access_token" in token
+    assert "token_type" in token
+
+
+def test_get_token_with_invalid_user(client, user):
+    response = client.post(
+        "/token",
+        data={"username": "usuario_inexistente", "password": "senha_inexistente"},
+    )
+
+    assert response.status_code == HTTPStatus.BAD_REQUEST
+
+
+def test_get_token_with_invalid_password(client, user):
+    response = client.post(
+        "/token",
+        data={"username": user.email, "password": "senha_inexistente"},
+    )
+
+    assert response.status_code == HTTPStatus.BAD_REQUEST
+
+
 def test_root_deve_retornar_ok_e_ola_mundo(client):
     # client = TestClient(app)  # Arrange
 
@@ -84,9 +114,10 @@ def test_read_user_with_user(client, user):
     assert response.json() == user_schema
 
 
-def test_update_user(client, user):
+def test_update_user(client, user, token):
     response = client.put(
-        "/users/1",
+        f"/users/{user.id}",
+        headers={"Authorization": f"Bearer {token}"},
         json={
             "username": "bob",
             "email": "bob@example.com",
@@ -97,29 +128,52 @@ def test_update_user(client, user):
     assert response.json() == {
         "username": "bob",
         "email": "bob@example.com",
-        "id": 1,
+        "id": user.id,
     }
 
 
-def test_update_non_existent_user(client):
+def test_update_wrong_user_id(client, user, token):
     response = client.put(
-        "/users/2",
+        f"/users/{user.id + 1}",
+        headers={"Authorization": f"Bearer {token}"},
         json={
             "username": "bob",
             "email": "bob@example.com",
             "password": "mynewpassword",
         },
     )
-    assert response.status_code == HTTPStatus.NOT_FOUND
+    assert response.status_code == HTTPStatus.BAD_REQUEST
 
 
-def test_delete_user(client, user):
-    response = client.delete("/users/1")
+# TESTE PRÉ TOKEN JWT
+# def test_update_non_existent_user(client):
+#     response = client.put(
+#         "/users/2",
+#         json={
+#             "username": "bob",
+#             "email": "bob@example.com",
+#             "password": "mynewpassword",
+#         },
+#     )
+#     assert response.status_code == HTTPStatus.NOT_FOUND
+
+
+def test_delete_user(client, user, token):
+    response = client.delete(
+        f"/users/{user.id}",
+        headers={"Authorization": f"Bearer {token}"},
+    )
     assert response.status_code == HTTPStatus.OK
     assert response.json() == {"message": "User deleted"}
 
 
-def test_delete_non_existent_user(client):
-    response = client.delete("/users/1")
+def test_delete_wrong_user_id(client, user, token):
+    response = client.delete(f"/users/{user.id + 1}", headers={"Authorization": f"Bearer {token}"})
+    assert response.status_code == HTTPStatus.BAD_REQUEST
 
-    assert response.status_code == HTTPStatus.NOT_FOUND
+
+# TESTE PRÉ TOKEN JWT
+# def test_delete_non_existent_user(client):
+#     response = client.delete("/users/1")
+
+#     assert response.status_code == HTTPStatus.NOT_FOUND
